+++ lib/redis/time_series/client.rb
@@ -42,9 +42,9 @@ class Redis
         @debug = !!bool
       end
 
-      # @return [Redis] the current Redis client. Defaults to +Redis.current+
+      # @return [Redis] the current Redis client. Defaults to +Redis.new+
       def redis
-        @redis ||= Redis.current
+        @redis ||= Redis.new
       end
 
       # Set the default Redis client for time series objects.
@@ -68,7 +68,7 @@ class Redis
       end
 
       def cmd_with_redis(redis, name, *args)
-        args = args.flatten.compact.map { |arg| arg.is_a?(Time) ? arg.ts_msec : arg }
+        args = args.flatten.compact.map { |arg| arg.is_a?(Time) ? arg.ts_msec : arg.to_s }
         puts "DEBUG: #{name} #{args.join(' ')}" if debug
         redis.call name, args
       end